removed console logs and renamed townhall to townhallmiddleware
